+++ tests/lib.js
@@ -1,0 +1,5 @@
+'use strict';
+
+describe('lib', () => {
+  require('./shared').runTests(require('../index'));
+});